+++ packages/client/src/core/GameLogic/getLevel.ts
@@ -3,6 +3,26 @@ import levelConfig, {
   Levels,
 } from './levelsConfig';
 
+export type LevelListType = {
+  title: string;
+  id: Levels;
+};
+
+export const levelList: LevelListType[] = [
+  {
+    title: 'Первый',
+    id: 'first',
+  },
+  {
+    title: 'Второй',
+    id: 'second',
+  },
+  {
+    title: 'Третий',
+    id: 'third',
+  },
+];
+
 export const getLevel = (
   context: CanvasRenderingContext2D,
   level: Levels,
