+++ packages/client/src/pages/End/End.tsx
@@ -10,7 +10,7 @@ export default function End() {
   const navigate = useNavigate();
   const { state } = useLocation();
   const handleRetry = useCallback(() => {
-    navigate(ROUTES.Game.path, {
+    navigate(ROUTES.Start.path, {
       state: { level: state.level },
     });
   }, [navigate]);
